Update 'video_pose_detection_yolo' example
- Use YOLOv11 instead of YOLOv8
- Use DataChain Yolo models
