+++ .gitignore
@@ -1,3 +1,10 @@
+# Mac OS X
+.DS_Store
+
+# IDE
+.idea
+.vscode
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
