+++ computer_vision/video_pose_detection_yolo/.gitignore
@@ -1,0 +1,5 @@
+.datachain
+data
+dev
+test-notebooks.sh
+*.pt
