+++ computer_vision/video_pose_detection_yolo/requirements.txt
@@ -1,0 +1,4 @@
+datachain
+imageio[ffmpeg]
+opencv-python
+ultralytics
