--- computer_vision/video_pose_detection_yolov8/.gitignore
@@ -1,5 +1,0 @@
-.datachain
-data
-dev
-test-notebooks.sh
-yolov8n-pose.pt
