--- computer_vision/video_pose_detection_yolov8/README.md
@@ -1,47 +1,0 @@
-<img src="https://raw.githubusercontent.com/iterative/datachain/main/docs/assets/datachain-sys-theme.svg" alt="Dataset" style="width: 200px;"/>
-
-# Tutorial: Enhancing Video Dataset Management and Frame-Level Annotation for YOLOv8 Pose Detection projects
-
-This tutorial dives into techniques to manage video and image datasets for Video Analytics and Vision AI projects. 
-
-📋 Topics covered:
-1. Building a Video DataChain for `kinetics-700-2020` video dataset
-2. Creating a Data Model for YOLOv8 Pose Detection projects
-3. Integrating Video-Level Annotations from CSV
-4. Extract and Manage Video Frames
-5. Running Pose Detection with YOLOv8 and Saving to DataChain
-6. Visualizing Pose Data
-
-
-## 🛠️ Install
-
-```bash
-python -m venv .venv
-source .venv/bin/activate
-pip install -r requirements.txt
-```
-
-You might also need to run:
-
-```
-apt-get update
-apt-get install -y --no-install-recommends libgl1 libglib2.0-0
-```
-
-## 🚀 Run Jupyter Notebooks
-
-The tutorial is available in Jupyter Notebooks. Start Jupyter Notebook server and follow the instructions.
-
-```bash
-jupyter notebook
-```
-
-## 🤝 Get Involved
-
-We'd love to have you join our growing community of DataChain users and contributors! Here's how you can get involved:
-
-- ⭐ Give us a star on [GitHub](https://github.com/iterative/datachain) to show your support
-- 🌐 Visit the [datachain.ai](https://datachain.ai) website to learn more about products and services
-- 🙌 Follow us on [LinkedIn](https://www.linkedin.com/company/dvc-ai/) and [Twitter](https://x.com/DVCorg) for the latest updates and insights
-
-Thanks for choosing DataChain, and happy coding! 😄
