--- computer_vision/video_pose_detection_yolov8/requirements.txt
@@ -1,2 +1,0 @@
-datachain
-ultralytics
